+++ app/static/index.html
@@ -16,9 +16,9 @@
     <script src='https://cdnjs.cloudflare.com/ajax/libs/leaflet/0.7.3/leaflet.js'></script>
     <script src='https://api.tiles.mapbox.com/mapbox.js/plugins/leaflet-markercluster/v0.4.0/leaflet.markercluster.js'></script>
     <script src='https://api.tiles.mapbox.com/mapbox.js/plugins/leaflet-locatecontrol/v0.43.0/L.Control.Locate.min.js'></script>
-    <script src='js/lib/angular-leaflet-directive.min.js'></script>
-    <script src='js/lib/pinchzoom.min.js'></script>
-    <script src='js/app.min.js'></script>
+    <script src='lib/angular-leaflet-directive.min.js'></script>
+    <script src='lib/pinchzoom.min.js'></script>
+    <script src='build/app.min.js'></script>
 
 
     <link rel='stylesheet' href='https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/css/bootstrap.min.css'>
@@ -28,7 +28,7 @@
     <link rel='stylesheet' href='https://api.tiles.mapbox.com/mapbox.js/plugins/leaflet-locatecontrol/v0.43.0/L.Control.Locate.mapbox.css'>
     <link rel='stylesheet' href='https://api.tiles.mapbox.com/mapbox.js/plugins/leaflet-locatecontrol/v0.43.0/css/font-awesome.min.css'>
 
-    <link rel='stylesheet' href='css/app.min.css'>
+    <link rel='stylesheet' href='build/app.min.css'>
     <link rel='icon' type='image/png' href='img/cgraphico.ico'>
 
     <link rel='apple-touch-icon' href='img/apple-touch-icon.png'>
