+++ app/static/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "civic-graph",
+  "version": "0.0.1",
+  "devDependencies": {
+    "gulp": "^3.9.1",
+    "gulp-concat": "^2.6.0",
+    "gulp-notify": "^2.2.0",
+    "gulp-uglify": "^1.5.4",
+    "gulp-clean-css": "^2.0.10"
+  }
+}
