+++ checkin.js
